+++ pkg/helper/mount_windows.go
@@ -115,7 +115,8 @@ func GetMountedFilePathWithBasePath(basePath, filePath string) string {
 	}
 	colonPos := strings.Index(filePath, ":")
 	if -1 == colonPos {
-		return ""
+		// Case: containerd, filePath is not start with driver letter. e.g. \\var\\logs\\pods\\xxx\\0.log
+		return filePath
 	}
 	return basePath + filePath[colonPos+1:]
 }
